Detect recursive lock between yp_all() and do_ypcall()
If the system is configured in such a way, that NIS is required
to contact the NIS server (most likely using a FQDN for the NIS
server and NIS for resolving host names), a call of yp_all() will
end in a call to do_ypcall(), but since yp_al() holds already the
lock, do_ypcall() will wait forever for the lock.
If we detect this, bail out with an error.