Detect recursive NIS calls
If the system is configured in such a way, that NIS is required
to contact the NIS server (most likely using a FQDN for the NIS
server and NIS for resolving host names), a call of do_ypcall()
and yp_all() can call themself again. This will lead to an
infinite recursion with a stack overflow crash.
If we detect this, bail out with an error.